pyproject.toml
[project]
name = "paperwork-engine"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.3",
    "dature[yaml]>=0.16.0",
    "jinja2>=3.1.6",
    "mdformat-obsidian[recommended]>=0.1.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
    "transliterate>=1.10.2",
]

[project.scripts]
paperwork = "paperwork.cli:main"

[build-system]
requires = ["setuptools >= 80.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["paperwork"]
